Make installation check for existing installation

+++ zsh/install.sh
@@ -1,5 +1,11 @@
-wget --output-document ~/.antigen.zsh https://raw.githubusercontent.com/zsh-users/antigen/master/antigen.zsh
+if [ ! -e ~/.antigen.zsh ]
+then
+    wget --output-document ~/.antigen.zsh https://raw.githubusercontent.com/zsh-users/antigen/master/antigen.zsh
+fi
 
-wget --output-document ~/.fasd.zsh https://raw.githubusercontent.com/clvv/fasd/master/fasd
-chmod 755 ~/.fasd.zsh
-ln -s ~/.fasd.zsh ~/.local/bin/fasd
+if [ ! -e ~/.fasd.zsh ]
+then
+    wget --output-document ~/.fasd.zsh https://raw.githubusercontent.com/clvv/fasd/master/fasd
+    chmod 755 ~/.fasd.zsh
+    ln -s ~/.fasd.zsh ~/.local/bin/fasd
+fi
